Proxy server that authenticates websocket connections against keystone.
See README.md for details on how the proxy works, and its usage.

This still needs work (automated test, RPM), but sending out now to
get early feedback.

The proxy is written in node.js because it is the ideal stack for
this kind of server. I'll explain this in more detail my coming
tech talk. In the meantime, I'd like Sirish and Bob to review this.
Bob has less experience with Javascript, but hopefully the overall
structure of the program should be clear.

This has been manually tested.

TODOs:
- restructure for unit testing
- RPM
- Potentially authorize based on user's role. Right now any
  valid token allows the client to connect to the destination.