Make pf9-authproxy support multiple cookies (IAAS-369) + RPM enhancements
Madhura found a case whereby after browsing Horizon on the DU,
multiple cookies were set and sent for the DU domain.
Also improved the RPM to clean up service upon package uninstallation.

Testing: updated unit test. Installed and tested RPM on
pf9.platform9.net DU. Browsed to Horizon, back to pf9 UI, confirmed
that multiple cookies were sent to the domain, and that pf9-authproxy
keystone authentication was working. Uninstalled RPM and verified
that service process was killed, and service no longer registered
in chkconfig.

Note: with this change, we now support cookie values that are
URL encoded, a feature of the 'cookie' library.
Since a keystone token can contain '=' characters,
it should be URL-encoded because '=' is also significant in
cookie strings. The new test data includes both encoded and
non-encoded versions of a token that contains '='. However,
due to a quirk in the cookie parsing library (and possibly the
cookie spec itself), the non-encoded token is also parsed
correctly, due to the parsing algorithm scanning forward to
the next ';' separator or the end of the cookie string.
I may still modify the UI code to URL encode the token regardless,
just in case.